Augment subject when making LogControl requests
So in the LogControlServlet, when a post or a get happens, the auth
data and AuthenticatorImpl are used, but by passing in false to
getSubject, the subject is not augmented.

For LSST, we start off with a token credential, then when we augment
this we pull out data such as the username, userid, etc.  So to have
the username to check against, the subject must be augmented.